Add Avenue
Also adds a new Web3 category as this is a new (and growing) industry.
It isn't necessarily related to fintech (e.g. Avenue isn't) as a lot of
companies in the space aren't. There are also quite a few opening up in
Lisbon so it makes sense to make this compatible and open a category of
its own.